+++ src/main/java/com/voshodnerd/BeatySalon/model/dto/ServiceItemDTO.java
@@ -9,6 +9,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToMany;
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -17,7 +18,7 @@ import java.util.Set;
 public class ServiceItemDTO {
     Long id;
     String name;
-    Integer price;
+    BigDecimal price;
     Integer durationInMinute;
     String category;
     Boolean active;
